pyproject.toml
[tool.poetry]
name = "configman"
version = "0.1.0"
description = ""
authors = ["Maurizio Sambati <maurizio@skicelab.com>"]

[tool.poetry.dependencies]
python = "^3.10"
apischema = "^0.17.5"
pytest = "^7.1.3"
PyYAML = {version = "^6.0", optional = true}
tomli = {version = "^2.0.1", optional = true}

[tool.poetry.group.dev.dependencies]
types-PyYAML = "^6.0.11"
isort = {version = "^5.10.1", extras = ["toml"]}
mypy = "^0.971"
flake8 = "^5.0.4"
black = "^22.6.0"
pre-commit = "^2.20.0"
ipython = "^8.4.0"

[tool.poetry.extras]
yaml = ["PyYAML"]
toml = ["tomli"]
all = ["PyYAML", "tomli"]

[tool.isort]
known_first_party = "configman"
profile = "black"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
